refactor: 폴더명 변경 및 폴더 이동
